Delegate DOM event handlers to common proxy handler.
- Other:
  - Rename (again) createNode→h. hyperapp/hyperapp/issues/632
