+++ src/h.js
@@ -1,0 +1,30 @@
+export function h(name, attributes /*...rest*/) {
+  var rest = []
+  var children = []
+  var length = arguments.length
+
+  while (length-- > 2) rest.push(arguments[length])
+
+  while (rest.length) {
+    var node = rest.pop()
+    if (
+      node &&
+      node.pop // isArray
+    ) {
+      for (length = node.length; length--; ) {
+        rest.push(node[length])
+      }
+    } else if (node != null && node !== true && node !== false) {
+      children.push(node)
+    }
+  }
+
+  return typeof name === "function"
+    ? name(attributes || {}, children) // h(Component)
+    : {
+        nodeName: name,
+        attributes: attributes || {},
+        children: children,
+        key: attributes && attributes.key
+      }
+}
