+++ test/h.test.js
@@ -1,7 +1,7 @@
-import { createNode as U } from "../src"
+import { h } from "../src"
 
 test("empty vnode", () => {
-  expect(U("div")).toEqual({
+  expect(h("div")).toEqual({
     nodeName: "div",
     attributes: {},
     children: []
@@ -9,13 +9,13 @@ test("empty vnode", () => {
 })
 
 test("vnode with a single child", () => {
-  expect(U("div", {}, ["foo"])).toEqual({
+  expect(h("div", {}, ["foo"])).toEqual({
     nodeName: "div",
     attributes: {},
     children: ["foo"]
   })
 
-  expect(U("div", {}, "foo")).toEqual({
+  expect(h("div", {}, "foo")).toEqual({
     nodeName: "div",
     attributes: {},
     children: ["foo"]
@@ -23,19 +23,19 @@ test("vnode with a single child", () => {
 })
 
 test("positional String/Number children", () => {
-  expect(U("div", {}, "foo", "bar", "baz")).toEqual({
+  expect(h("div", {}, "foo", "bar", "baz")).toEqual({
     nodeName: "div",
     attributes: {},
     children: ["foo", "bar", "baz"]
   })
 
-  expect(U("div", {}, 0, "foo", 1, "baz", 2)).toEqual({
+  expect(h("div", {}, 0, "foo", 1, "baz", 2)).toEqual({
     nodeName: "div",
     attributes: {},
     children: [0, "foo", 1, "baz", 2]
   })
 
-  expect(U("div", {}, "foo", U("div", {}, "bar"), "baz", "quux")).toEqual({
+  expect(h("div", {}, "foo", h("div", {}, "bar"), "baz", "quux")).toEqual({
     nodeName: "div",
     attributes: {},
     children: [
@@ -60,7 +60,7 @@ test("vnode with attributes", () => {
     }
   }
 
-  expect(U("div", attributes, "baz")).toEqual({
+  expect(h("div", attributes, "baz")).toEqual({
     nodeName: "div",
     attributes,
     children: ["baz"]
@@ -74,21 +74,21 @@ test("skip null and Boolean children", () => {
     children: []
   }
 
-  expect(U("div", {}, true)).toEqual(expected)
-  expect(U("div", {}, false)).toEqual(expected)
-  expect(U("div", {}, null)).toEqual(expected)
+  expect(h("div", {}, true)).toEqual(expected)
+  expect(h("div", {}, false)).toEqual(expected)
+  expect(h("div", {}, null)).toEqual(expected)
 })
 
 test("components", () => {
-  const Component = (attributes, children) => U("div", attributes, children)
+  const Component = (attributes, children) => h("div", attributes, children)
 
-  expect(U(Component, { id: "foo" }, "bar")).toEqual({
+  expect(h(Component, { id: "foo" }, "bar")).toEqual({
     nodeName: "div",
     attributes: { id: "foo" },
     children: ["bar"]
   })
 
-  expect(U(Component, { id: "foo" }, [U(Component, { id: "bar" })])).toEqual({
+  expect(h(Component, { id: "foo" }, [h(Component, { id: "bar" })])).toEqual({
     nodeName: "div",
     attributes: { id: "foo" },
     children: [
@@ -103,9 +103,9 @@ test("components", () => {
 
 test("component with no attributes adds default attributes", () => {
   const Component = ({ name = "world" }, children) =>
-    U("div", {}, "Hello " + name)
+    h("div", {}, "Hello " + name)
 
-  expect(U(Component)).toEqual({
+  expect(h(Component)).toEqual({
     nodeName: "div",
     attributes: {},
     children: ["Hello world"]
